refactor: move map
